refactor(blog): consolidate client components into page files
- Migrate BlogClient and BlogsClient functionality directly into page components
- Convert blog pages to client components with useEffect for data fetching
- Remove separate client component files to simplify architecture
- Add new layout files for blog routes
- Maintain all existing functionality including SEO metadata and structured data

This refactoring consolidates the blog feature into self-contained page components, reducing file complexity and improving maintainability while preserving all SEO and rendering capabilities.
